+++ EstimoteSDKv5/EstimoteSDKv5.framework/Versions/A/Headers/EstimoteSDKv5.h
@@ -7,7 +7,7 @@
 //  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
 //
 //
-//  Version: 5.0.0-alpha.4
+//  Version: 5.0.0-alpha.5
 //  Copyright (c) 2017 Estimote. All rights reserved.
 
 #import "ESTCloudCredentials.h"