+++ Examples/swift/DeskObserver/DeskObserver/AppDelegate.swift
@@ -12,10 +12,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        // TODO: put your App ID and App Token here
-        // You can get them by adding your app on https://cloud.estimote.com/#/apps
-        ESTCloudCredentials.setUpSharedInstance(appID: "<#App ID#>", appToken: "<#App Token#>")
-        
         return true
     }